CSS
Dumped all CSS into git. formatting for the desktop view is still incomplete. Currently unable to align cards in the desired formatting. I have attempted to utilize grid and flex with varying degrees of success, but none could quite get the formatting where it needed to be.  Note that the green and blue spaces in the desktop view are there to help me visualize my divs, they will be removed once I successfully format that page.